+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/error/ExtensionCircularDependencyErrorImpl.java
@@ -19,10 +19,11 @@ import org.apache.commons.lang3.builder.ToStringStyle;
 import tools.jackson.databind.annotation.*;
 
 /**
- *  <p>Returned when a referenced Extension does not exist or is not applicable to the same trigger.</p>
+ *  <p>Returned when a circular reference is detected among Extension dependencies.</p>
+ *  <p>The <code>message</code> lists the <code>id</code> of each Extension in the detected cycle, starting and ending with the same Extension.</p>
  */
 @Generated(value = "io.vrap.rmf.codegen.rendering.CoreCodeGenerator", comments = "https://github.com/commercetools/rmf-codegen")
-public class MissingDependencyErrorImpl implements MissingDependencyError, ModelBase {
+public class ExtensionCircularDependencyErrorImpl implements ExtensionCircularDependencyError, ModelBase {
 
     private String code;
 
@@ -34,18 +35,18 @@ public class MissingDependencyErrorImpl implements MissingDependencyError, Model
      * create instance with all properties
      */
     @JsonCreator
-    MissingDependencyErrorImpl(@JsonProperty("message") final String message,
+    ExtensionCircularDependencyErrorImpl(@JsonProperty("message") final String message,
             @JsonAnySetter @JsonProperty("values") final Map<String, java.lang.Object> values) {
         this.message = message;
         this.values = values;
-        this.code = MISSING_DEPENDENCY;
+        this.code = EXTENSION_CIRCULAR_DEPENDENCY;
     }
 
     /**
      * create empty instance
      */
-    public MissingDependencyErrorImpl() {
-        this.code = MISSING_DEPENDENCY;
+    public ExtensionCircularDependencyErrorImpl() {
+        this.code = EXTENSION_CIRCULAR_DEPENDENCY;
     }
 
     /**
@@ -57,7 +58,7 @@ public class MissingDependencyErrorImpl implements MissingDependencyError, Model
     }
 
     /**
-     *  <p><code>"The extensions '[ext-1, ext-2]' referenced in 'dependencies' do not exist."</code></p>
+     *  <p><code>"Circular dependency detected: [3fa85f64-5717-4562-b3fc-2c963f66afa6, 5e0e1e0b-1f4b-4b8b-9c1a-2d3f4a5b6c7d, 3fa85f64-5717-4562-b3fc-2c963f66afa6]"</code></p>
      */
 
     public String getMessage() {
@@ -91,7 +92,7 @@ public class MissingDependencyErrorImpl implements MissingDependencyError, Model
         if (o == null || getClass() != o.getClass())
             return false;
 
-        MissingDependencyErrorImpl that = (MissingDependencyErrorImpl) o;
+        ExtensionCircularDependencyErrorImpl that = (ExtensionCircularDependencyErrorImpl) o;
 
         return new EqualsBuilder().append(code, that.code)
                 .append(message, that.message)
@@ -116,7 +117,7 @@ public class MissingDependencyErrorImpl implements MissingDependencyError, Model
     }
 
     @Override
-    public MissingDependencyError copyDeep() {
-        return MissingDependencyError.deepCopy(this);
+    public ExtensionCircularDependencyError copyDeep() {
+        return ExtensionCircularDependencyError.deepCopy(this);
     }
 }